+++ src/main/java/org/openstreetmap/osmium/data/BuildingImport.java
@@ -8,7 +8,7 @@ public class BuildingImport extends AbstractImport {
 
     Integer levels;
 
-    Integer height;
+    Float height;
 
     Integer area;
 
@@ -45,11 +45,11 @@ public class BuildingImport extends AbstractImport {
         this.levels = levels;
     }
 
-    public Integer getHeight() {
+    public Float getHeight() {
         return height;
     }
 
-    public void setHeight(Integer height) {
+    public void setHeight(Float height) {
         this.height = height;
     }
